requirements.txt
Flask==3.1.3
python-dotenv==1.2.2
groq==1.1.2
chromadb==0.5.21
sentence-transformers==3.3.1
pypdf==5.1.0